requirements.txt
pandas
numpy
scipy
scikit-learn
imblearn
xgboost
lightgbm
seaborn
matplotlib
torch
torchvision
pillow
opencv-python
h5py